requirements.txt
# Core app
streamlit
pandas
numpy
scikit-learn

# Notebook + plots
jupyter
matplotlib
seaborn

# CV parsing
python-docx
pypdf
